move hash_password and check_password_hash to webapp.common.api.utils
