+++ src/eduid/webapp/reset_password/helpers.py
@@ -36,7 +36,6 @@ from dataclasses import dataclass
 from enum import unique
 from typing import Any, Dict, List, Mapping, Optional, Union
 
-import bcrypt
 from flask import render_template
 from flask_babel import gettext as _
 
@@ -55,7 +54,7 @@ from eduid.userdb.user import User
 from eduid.webapp.common.api.exceptions import MailTaskFailed, ThrottledException
 from eduid.webapp.common.api.helpers import send_mail
 from eduid.webapp.common.api.messages import FluxData, TranslatableMsg, error_response, success_response
-from eduid.webapp.common.api.utils import get_short_hash, get_unique_hash, save_and_sync_user
+from eduid.webapp.common.api.utils import check_password_hash, get_short_hash, get_unique_hash, save_and_sync_user
 from eduid.webapp.common.api.validation import is_valid_password
 from eduid.webapp.common.authn import fido_tokens
 from eduid.webapp.common.authn.utils import generate_password
@@ -178,7 +177,7 @@ def get_pwreset_state(email_code: str) -> Union[ResetPasswordEmailState, ResetPa
 
 
 def is_generated_password(password: str) -> bool:
-    if check_password(password, session.reset_password.generated_password_hash):
+    if check_password_hash(password, session.reset_password.generated_password_hash):
         current_app.logger.info('Generated password used')
         current_app.stats.count(name='generated_password_used')
         return True
@@ -244,26 +243,6 @@ def generate_suggested_password(password_length: int) -> str:
     password = ' '.join([password[i * 4 : i * 4 + 4] for i in range(0, math.ceil(len(password) / 4))])
 
     return password
-
-
-def hash_password(password: str) -> str:
-    """
-    Return a hash of the provided password
-
-    :param password: password as plaintext
-    """
-    password = ''.join(password.split())
-    return bcrypt.hashpw(password, bcrypt.gensalt())
-
-
-def check_password(password: str, hashed: Optional[str]) -> bool:
-    """
-    Check that the provided password corresponds to the provided hash
-    """
-    if hashed is None:
-        return False
-    password = ''.join(password.split())
-    return bcrypt.checkpw(password, hashed)
 
 
 def extra_security_used(state: ResetPasswordState, security_key_used: bool = False) -> bool:
